Add Alt 0 header.
